part2 finished

+++ malloc_1.cpp
@@ -2,7 +2,7 @@
 #include <unistd.h>
 
 void* smalloc(size_t size){
-    if(size <= 0 || size > pow(10,8)){
+    if(size <= 0 || size > size_t(pow(10.0,8.0))){
         return nullptr;
     }
     void* ptr = sbrk(size);
